requirements.txt
requests
python-telegram-bot
schedule
python-dotenv
selenium
webdriver_manager


